Rename backbone config
