+++ misc/pytorch_toolkit/text_recognition/configs/handwritten-english-recognition-0001.yml
@@ -1,5 +1,5 @@
 backbone_config:
-  type: cnn
+  type: hetr
   output_channels: 80
 head:
   type: LSTMEncoderDecoder
