+++ misc/pytorch_toolkit/text_recognition/text_recognition/models/model.py
@@ -20,7 +20,7 @@ import torch
 import torch.nn as nn
 
 from .backbones.resnet import (CustomResNetLikeBackbone, ResNetLikeBackbone)
-from .backbones.cnn import HETRBackbone
+from .backbones.hetr import HETRBackbone
 from .text_recognition_heads.attention_based import AttentionBasedLSTM
 from .text_recognition_heads.attention_based_2d import \
     TextRecognitionHeadAttention
@@ -36,7 +36,7 @@ TEXT_REC_HEADS = {
 BACKBONES = {
     'resnet': ResNetLikeBackbone,
     'custom_resnet': CustomResNetLikeBackbone,
-    'cnn': HETRBackbone,
+    'hetr': HETRBackbone,
 }
 
 TRANSFORMATIONS = {
